Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -2,7 +2,7 @@
 
 int main()
 {
-    float num1,num2,num3;
+    int num1,num2,num3;
     float avg;
     scanf("%d%d%d",&num1,&num2,&num3);
     avg=(num1+num2+num3)/3;
